Fix package names
This commit modifies CMakeLists.txt and package.xml so that the
package name ends with "_msgs". The package names are also renamed so
that the same package name is used for both ROS1 and ROS2.

This is the easiest way to bridge the ROS1 and ROS2 messages.

See https://github.com/ros2/ros1_bridge/blob/master/doc/index.rst